fix(frontend): move COUNT_BADGE_SX out of the token definition file
styles/theme-constants.ts is where the hardcoded-styles CI check finds every
literal color/spacing value everywhere else in the app -- it's the token
*definition* site, so editing it drags its own pre-existing literals (the
GREYSCALE hexes, BACKDROP_COLORS, ELEVATION) into the diff the checker scans.

COUNT_BADGE_SX is nav chrome, not a design token, so it belongs in
sidebar-utils.ts beside collapsedNavItemSx. theme-constants.ts and theme.ts
now match main exactly.
